requirements.txt
asgiref==3.6.0
attrs==23.1.0
Django==3.1.14
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.11.2
factory-boy==2.12.0
Faker==18.4.0
more-itertools==9.1.0
packaging==23.1
pluggy==0.13.1
psycopg2-binary==2.9.6
py==1.11.0
pytest==5.4.3
pytest-django==3.9.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
wcwidth==0.2.6
